refactor: extract parsing utilities and modularise components
